chore: streamline dialog components namespace exports
